// src/app/items/model/allowed-operators.enum.ts
export enum AllowedOperators {
  EQUALS = 'equals',
  NOT_EQUALS = 'not equals',
  CONTAINS = 'contains',
  DOES_NOT_CONTAIN = 'does not contain',
  LESS_THAN = 'less than',
  GREATER_THAN = 'greater than',
  LESS_THAN_OR_EQUAL_TO = 'less than equal to',
  GREATER_THAN_OR_EQUAL_TO = 'greater than equal to',
}